sobun/requirements.txt
# This requirements file sets up the virtual environment for python 
-i https://pypi.org/simple
amqp==2.3.2
billiard==3.5.0.4
bluepy==1.2.0; sys_platform=="linux2"
celery==4.2.1
docopt==0.6.2
google-cloud-bigquery==1.6.0
google-cloud-pubsub==0.38.0
kombu==4.2.1
pytz==2018.5
redis==2.10.6
vine==1.1.4
